requirements.txt
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
python-telegram-bot==22.3
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
